pyproject.toml
[build-system]
requires = ["maturin>=1.8.0,<2.0"]
build-backend = "maturin"

[tool.maturin]
python-source = "src"
module-name = "patchon._native._rust_ext"
manifest-path = "rust/Cargo.toml"
features = ["pyo3/extension-module"]

[project]
name = "patchon"
version = "0.1.0"
description = "Run Python scripts with temporary source-file hot patches applied before execution and restored afterward."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "patchon contributors" }
]
keywords = ["python", "patch", "cli", "hotfix", "development"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Build Tools",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Utilities",
]
dependencies = [
  "PyYAML>=6.0,<7.0",
]

[project.urls]
Homepage = "https://github.com/cskujou/patchon"
Repository = "https://github.com/cskujou/patchon"
Issues = "https://github.com/cskujou/patchon/issues"
Documentation = "https://github.com/cskujou/patchon"

[project.scripts]
patchon = "patchon.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-cov>=5.0,<6.0",
  "ruff>=0.8.0",
  "mypy>=1.12.0",
  "maturin>=1.8.0",
  "mkdocs>=1.6.0",
  "mkdocs-material>=9.5.0",
  "mkdocs-click>=0.8.0",
  "mkdocstrings-python>=1.13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/patchon"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "F",   # Pyflakes
  "I",   # isort
  "N",   # pep8-naming
  "W",   # pycodestyle warnings
  "UP",  # pyupgrade
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "SIM", # flake8-simplify
]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
show_error_codes = true
